// app/Interfaces/ISec_Users.ts
export interface ISec_Users {
    UserName: string;
    password: string;
    isActive: boolean;
    roleId: number;
    mobile: string;
    email: string;
    address: string;
    tel: string;
    Isdeleted: boolean;
    FullName: string;
    Department: number;
    image: string;
    userId: number;
    guid:string;
    DepartmentName:string;
    RoleName:string;
}